+++ lib/astro_chart.rb
@@ -20,6 +20,8 @@ require_relative "astro_chart/progressions"
 require_relative "astro_chart/composite"
 require_relative "astro_chart/solar_return"
 require_relative "astro_chart/draconic"
+require_relative "astro_chart/dignities"
+require_relative "astro_chart/profection"
 
 module AstroChart
 end